refactor: 카테고리, 등록폼 타입 폴더 구조 변경 및 import문 수정
