--- src/models/Account.ts
@@ -1,6 +1,0 @@
-export type CreateAccountForm = {
-  registerDate: string;
-  amount: number;
-  userCategoryId: number;
-  content?: string;
-};
