--- src/models/Category.ts
@@ -1,7 +1,0 @@
-export type CategoryType = 'income' | 'expenditure';
-
-export type Category = {
-  id: number;
-  name: string;
-  categoryType: CategoryType;
-};
