--- src/models/index.ts
@@ -1,4 +1,0 @@
-import { Category } from './Category';
-import { CreateAccountForm } from './Account';
-
-export type { Category, CreateAccountForm };
