+++ src/pages/updateAccount/index.tsx
@@ -13,7 +13,7 @@ import {
   fetchDeleteIncomes,
   fetchDeleteExpenditures,
 } from '@api';
-import { CreateAccountForm } from '@models';
+import type { CreateAccountForm } from '@types';
 
 const ACCOUNT_TYPE: TabItem[] = [
   {
